See Commit Details
Changed the fonts. Roboto sometimes has a weird time rendering on some systems, so I changed it to plain old sans serif. For the title of posts, they are now Merriweather serif. The footer was cleaned up a bit as well, and I added a note linking back to this repository. I also added a date to the bottom right of posts on the home page so that at a glance you can tell when stuff was published.